+++ src/panoptes/pocs/state/machine.py
@@ -209,9 +209,7 @@ class PanStateMachine(Machine):
         state_changed = transition_method()
         if state_changed:
             self.logger.success(f"Finished with {self.state} state")
-            self.db.insert_current(
-                "state", {"source": self.state, "dest": self.next_state}, store_permanently=False
-            )
+            self.db.insert_current("state", {"source": self.state, "dest": self.next_state})
 
         return state_changed